adapt to coding style + move out roundrobin helper
